refactor: Move taskFromLine() out from Obsidian code to allow testing
